python/requirements.txt
fastapi[all]
cuid2
ffmpeg-python
opencv-python
numpy
imutils
transformers